app/requirements.txt
baidu-aip==2.2.18.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
ChineseTone==0.1.4
Click==7.0
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.15.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.0
oauthlib==3.1.0
pandas==0.25.3
Pillow==6.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
uritemplate==3.0.1
urllib3==1.25.7
Werkzeug==0.16.0
